+++ NpgsqlRest/UploadHandlers/UploadExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using NpgsqlRest.UploadHandlers.Handlers;
+using static NpgsqlRest.NpgsqlRestOptions;
 
 namespace NpgsqlRest.UploadHandlers;